\name{MapColoring-internal}
\alias{oc}
\alias{getNeighbors}
\alias{getAmountColor}
\title{
Internal Functions for Package MapColoring
}
\description{
Various internal functions.
}
\usage{
oc(cand.idx, coloring, adj_mat, cand.colors, beta)
getAmountColor(node_indexes, color_number, coloring)
getNeighbors(adj_mat, node_index)
}

\arguments{
}

\details{
}

\value{
}

\references{
}
\author{
Philipp Hunziker
}

\note{
}

\seealso{
}

\examples{
}

\keyword{internal}